requirements/requirements-base.txt
markdown==2.5.2
django-guardian==1.3.0
django-filter==0.10.0
djangorestframework==3.1.3
djangorestframework-jwt==1.4.0
django-rest-swagger==0.3.3
PyYAML==3.11
Pygments==2.0.2
